refactor: PinDetail 페이지에서 Update 관련 로직 분리
UpdatedPinDetail 페이지 컴포넌트로 분리한다.
